added utility functions for BlockNumber
