+++ objects/src/transaction/inputs.rs
@@ -44,10 +44,10 @@ impl TransactionInputs {
 
         // check the block_chain and block_header are consistent
         let block_num = block_header.block_num();
-        if block_chain.chain_length() != block_header.block_num().as_usize() {
+        if block_chain.chain_length() != block_header.block_num() {
             return Err(TransactionInputError::InconsistentChainLength {
-                expected: block_header.block_num().as_u32(),
-                actual: block_chain.chain_length() as u32,
+                expected: block_header.block_num(),
+                actual: block_chain.chain_length(),
             });
         }
